Added hwutils file
